fix signup redirect to otp form, enable login
